Add Redux Structure
- Set up the Redux store with hot module reloading
- Add a shared `App` component that's used for iOS and Android
- Set up a Redux structure using the [ducks-modular-redux](https://github.com/erikras/ducks-modular-redux) pattern.
- Add a simple counter module and wire it up to the UI to show how it works.

The Ducks pattern is my current preferred structure as it provides a nice
organization of your redux specific things in an isolated file so you aren't
passing having to constantly import constants, action creators etc.

It's also beneficial in that your transformations around a specific part of your
store in one place so it's easier to track down what's happening in your action
creators and reducers.

It supports growth well in that you can turn a single file module into
a directory and separate it out if it gets complicated, or break it up into
other modules while still keeping everything in fairly close together.