Make bin/generate portable to Linux
GNU sed parses its `-i` argument differently to the sed installed on
macOS. This change introduces a `sed_in_place` function that handles the
difference, and switches to `bash` so that we can export the function to
subshells, making it usable from within an invocation of `xargs`.

Includes a couple of other minor fixes to improve portability and
string quoting.