Fix incorrect env file being loaded on iOS
The buil pre-actions script was not loading build settings from the app target,
so it never had release set and was thus using the `.env.production` for all
builds.

This configures it to use build settings from the app target which provides that
env var and sets it up to load the `.local` variant for debug builds.