Dynamically set iOS build number
This number needs to be an ever-increasing integer. To enforce this, we
can automate the build number calculation by looking at how many commits
exist on master. This number will constantly increment with no
additional work on our end, and will give us a concrete anchor for what
the app code looked like for a specific release.