Improve package name generation
Previously, we were generating the package name fairly naively. Instead
of just plugging in the second part of the name blindly, we can instead
base it off our (new) company name variable, and attach a normalized
version of the project name as well. In addition, we'll let users
override the whole thing in general, which will let them override the
`com.` part of the name as well.

Finally, we'll use the package name in the iOS app as well as in the
Android app.